updating the go-pingdom dependency
Note: I'm not into go development, so I am not sure on how to update,
and I probably did it the wrong way (git clone, rm .git ...). But the
go get provided in the Makefile was not working for me.

Tests: runtime using:

```
$ ./pingdom-prometheus-exporter server dummy dummy <api key>
$ curl http://localhost:8000/metrics
[...]
```

Note: dummy/dummy because the API key is sufficient for authentication
now, user and password are unused anymore.